requirements.txt
torch
torchvision
pytorch_wavelets
numpy
scikit-learn
matplotlib
Pillow
tqdm
PyWavelets